# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Describe It - Spanish Learning API
  version: 1.0.0
  description: |
    A comprehensive API for visual-based Spanish learning through AI-powered image descriptions, 
    Q&A generation, vocabulary extraction, and translation services.
    
    This API enables users to:
    - Generate multi-style image descriptions in Spanish and English
    - Create interactive Q&A sessions for comprehension testing
    - Extract and translate vocabulary from descriptions
    - Track learning progress and export data
    - Search and manage images for learning content
    
  contact:
    name: Describe It Support
    email: support@describe-it.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://describe-it.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /export/generate:
    post:
      tags:
        - Export
      summary: Generate Export Data
      description: Export learning session data in various formats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Export generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExportRequest:
      type: object
      required:
        - sessionId
        - format
      properties:
        sessionId:
          type: string
          description: Session ID to export data for
          example: "session_123456789"
        format:
          type: string
          enum: ["pdf", "json", "csv"]
          description: Export format
          example: "pdf"
        includeImages:
          type: boolean
          default: true
          description: Whether to include images in export
          example: true
...
    ExportResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        downloadUrl:
          type: string
          format: uri
          example: "https://api.describe-it.app/exports/session_123456789.pdf"
        expiresAt:
          type: string
          format: date-time
          example: "2024-01-16T10:30:00Z"
...